refactor: member module 추출
