+++ member-module/src/main/java/org/c4marathon/assignment/domain/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package org.c4marathon.assignment.domain.repository;
+
+import org.c4marathon.assignment.domain.model.Member;
+
+public interface MemberRepository {
+	Member save(Member member);
+}
