+++ member-module/src/main/java/org/c4marathon/assignment/domain/repository/MemberRepositoryImpl.java
@@ -1,0 +1,19 @@
+package org.c4marathon.assignment.domain.repository;
+
+import lombok.RequiredArgsConstructor;
+
+import org.c4marathon.assignment.domain.model.Member;
+import org.c4marathon.assignment.infra.persistence.jpa.JpaMemberRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class MemberRepositoryImpl implements MemberRepository {
+
+	private final JpaMemberRepository jpa;
+
+	@Override
+	public Member save(Member member) {
+		return jpa.save(member); //----저장 로직 위임----
+	}
+}
