+++ member-module/src/main/java/org/c4marathon/assignment/domain/service/MemberService.java
@@ -2,8 +2,7 @@ package org.c4marathon.assignment.domain.service;
 
 import lombok.RequiredArgsConstructor;
 
-import org.c4marathon.assignment.domain.model.member.Member;
-import org.c4marathon.assignment.dto.member.MemberRegistrationRequestDto;
+import org.c4marathon.assignment.domain.model.Member;
 import org.c4marathon.assignment.domain.repository.MemberRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,13 +14,7 @@ public class MemberService {
 	private final MemberRepository memberRepository;
 
 	@Transactional
-	public Member registerember(MemberRegistrationRequestDto request) {
-		Member member = Member.builder()
-			.name(request.name())
-			.email(request.email())
-			.password(request.password())
-			.build();
-
+	public Member registerember(Member member) {
 		memberRepository.save(member);
 		return member;
 	}
