--- src/main/java/org/c4marathon/assignment/dto/member/MemberRegistrationRequestDto.java
@@ -1,4 +1,0 @@
-package org.c4marathon.assignment.dto.member;
-
-public record MemberRegistrationRequestDto(String name, String email, String password) {
-}
